requirements.txt
streamlit
fastapi
uvicorn
requests

pandas
numpy
tensorflow

langchain-core
langchain-groq
langchain-openai
langchain-google-genai
